Add nodes

+++ src/nodes/node_property.zig
@@ -64,15 +64,15 @@ pub fn drawAxisMaskProperty(self: *const NodeProperty, data: *[]u8) bool {
 }
 
 pub fn drawMaterialProperty(self: *const NodeProperty, data: *[]u8) bool {
-    var data_ptr: [*c]i32 = @ptrCast([*c]i32, @alignCast(@alignOf(i32), &data.*[self.offset]));
-    const current_index: i32 = data_ptr[0];
+    var data_ptr: [*c]usize = @ptrCast([*c]usize, @alignCast(@alignOf(usize), &data.*[self.offset]));
+    const current_index: usize = data_ptr[0];
 
     var changed: bool = false;
     if (nc.igBeginCombo(self.name.ptr, &Global.main_scene.materials.children.items[@intCast(usize, current_index)].name, nc.ImGuiComboFlags_None)) {
         for (Global.main_scene.materials.children.items) |mat, i| {
             nc.igPushID_Ptr(mat);
             if (nc.igSelectable_Bool(&mat.name, i == current_index, nc.ImGuiSelectableFlags_None, .{ .x = 0, .y = 0 })) {
-                data_ptr[0] = @intCast(i32, i);
+                data_ptr[0] = @intCast(usize, i);
                 changed = true;
             }
             nc.igPopID();
